-- (jabber-compat) added NonSASLAuthenticator that supports the old    jabber style authentication (newly tagged as iq-auth in the stream    features).  This authenticator supports both digest and plain. -- (jabber-compat) added AuthIQPacket. -- (jabber-compat) relocated VersionIQPacket to com.echomine.jabber.packet -- (xmpp) fixed another "auth"=>"stream" accidental string    replacement that was overlooked in the previous revision.
With this revision, by adding the jabber-compat library to the classpath,
old-style iq-auth authentication is automagically supported.  However,
current coding is not very relaxing on the parsing of old-style error
packets.  Thus, if you get an error packet during authentication, it
will likely use the old-style error stanzas, in which case the current
xmpp error packet unmarshaller will fail to understand and parse properly.
Hopefully this will be addressed in the future by adding a customized
error stanza packet mapper when jabber-compat library is added to the
classpath.  For now, assume that any exception thrown during login
is considered a failure in authentication.  The developer should retry
authentication with new credentials.